Выполняет 5ю дз
